feat: merge segmentation_and_annotation with general-example
