--- gen3/neural-networks/segmentation-and-detection/README.md
@@ -1,4 +1,0 @@
-General detection and segmentation recognition model.
-================
-
-This example demonstrates how one can utilize the new ImgDetctionsExtended class to create sgementation mask of the detected objects.
