--- gen3/neural-networks/segmentation-and-detection/requirements.txt
@@ -1,1 +1,0 @@
-depthai>=3.0.0-alpha.6
